Move generated columns query logic to column observer
